refactor: UserInput의 askInputUser() 메소드를 GameController로 옮김
- 사용자에게 물어보는 askInputUser()메소드는 Game에서 주관하는 로직이기 때문에 옮김
